volparametriceq - firt support for x86 - please test!

+++ plugins/audio_interface/volparametriceq/install.sh
@@ -3,12 +3,45 @@
 echo "Installing volparametriceq dependencies"
 echo "unload Loopback module if exists"
 sudo rmmod snd_aloop
-#sudo apt-get update
+#su[do apt-get update
 #sudo apt-get -y install caps
-sudo cp /data/plugins/audio_interface/volparametriceq/caps.tar /
-cd /
-sudo tar xvf caps.tar
-sudo rm /caps.tar
+libpath=/data/plugins/audio_interface/volparametriceq
+derrormess="Failed to extract caps"
+echo "Detecting cpu"
+cpu=$(lscpu | awk 'FNR == 1 {print $2}')
+
+if [ "$cpu" = "armv6l" ] || [ "$cpu" = "armv7l" ] || ["$cpu" = "aarch64" ];
+then
+	cd $libpath
+        echo "Cpu is $cpu, installing required caps version."
+	sudo cp /data/plugins/audio_interface/volparametriceq/caps.tar /
+	cd /
+	sudo tar xvf caps.tar
+	sudo rm /caps.tar
+	if [ $? -eq 0 ]
+		then
+			echo "Extracting data"
+		else
+			echo "$derrormess"
+			exit -1
+		fi
+elif [ $cpu = "i686" ]
+then
+cd $libpath
+        echo "Cpu is $cpu, installing required caps version."
+	sudo cp /data/plugins/audio_interface/volparametriceq/caps-x86.tar /
+	cd /
+	sudo tar xvf caps-x86.tar
+	sudo rm /caps-x86.tar
+	if [ $? -eq 0 ]
+		then
+			echo "Extracting data"
+		else
+			echo "$derrormess"
+			exit -1
+		fi
+fi
+
 if [ ! -f "/data/configuration/audio_interface/volparametriceq/config.json" ];
 	then
 		echo "file doesn't exist, nothing to do"
